split out useOnSelectChain, useSyncChainQuery
